+++ app/src/main/java/com/cmoney/backend2/sample/view/main/MainActivity.kt
@@ -108,6 +108,7 @@ class MainActivity : AppCompatActivity() {
 //                CMTalkServiceCase(),
 //                CommonServiceCase(this@MainActivity),
 //                CustomGroupServiceCase(),
+//                DataServiceCase(),
 //                DtnoServiceCase(),
 //                EmilyServiceCase(),
 //                ForumOceanServiceCase(),