+++ backend2/src/main/java/com/cmoney/backend2/data/service/api/FundIdData.kt
@@ -1,0 +1,6 @@
+package com.cmoney.backend2.data.service.api
+
+data class FundIdData(
+    val title: List<String>?,
+    val data: List<List<String>>?
+)